Fix: เพิ่มการส่งรายงานและรีวิว และเอารีวิวออกมาแสดง

+++ src/main/java/com/springboot/repository/NotificationRepository.java
@@ -3,13 +3,11 @@ package com.springboot.repository;
 import org.springframework.stereotype.Repository;
 import com.springboot.model.*;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
 @Repository
-public interface NotificationRepository extends JpaRepository<Notification, Long> {
+public interface NotificationRepository extends JpaRepository<Notification, Integer> {
     // List<Notification> findByRecipientIdAndIsReadFalse(Long userId);
 
     // 6371 = รัศมีโลก (กิโลเมตร)
